# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CruiseBooking",
        "title": "CruiseBooking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/admin/cruises": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cruise"
                ],
                "summary": "List cruises",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by company",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Keyword on cruise name",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by status, 0 means all",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by tonnage/name",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cruise"
                ],
                "summary": "Create cruise",
                "parameters": [
                    {
                        "description": "Cruise data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CruiseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/cruises/batch-status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cruise"
                ],
                "summary": "Batch update cruise status",
                "parameters": [
                    {
                        "description": "Batch status payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CruiseBatchStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/cruises/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cruise"
                ],
                "summary": "Update cruise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cruise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cruise data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CruiseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cruise"
                ],
                "summary": "Delete cruise",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cruise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cruisebooking_backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cruisebooking_backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码（0 表示成功）",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "internal_handler.CruiseBatchStatusRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "description": "需要更新状态的邮轮 ID 列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "目标状态",
                    "type": "integer"
                }
            }
        },
        "internal_handler.CruiseRequest": {
            "type": "object",
            "required": [
                "company_id",
                "name"
            ],
            "properties": {
                "build_year": {
                    "description": "建造年份",
                    "type": "integer"
                },
                "code": {
                    "description": "邮轮代码/英文简称",
                    "type": "string"
                },
                "company_id": {
                    "description": "所属公司 ID（必填）",
                    "type": "integer"
                },
                "crew_count": {
                    "description": "船员人数",
                    "type": "integer"
                },
                "deck_count": {
                    "description": "甲板层数",
                    "type": "integer"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "english_name": {
                    "description": "英文名称",
                    "type": "string"
                },
                "length": {
                    "description": "船体长度（米）",
                    "type": "number"
                },
                "name": {
                    "description": "邮轮名称（必填）",
                    "type": "string"
                },
                "passenger_capacity": {
                    "description": "最大载客量",
                    "type": "integer"
                },
                "refurbish_year": {
                    "description": "最近翻新年份",
                    "type": "integer"
                },
                "room_count": {
                    "description": "舱房总数",
                    "type": "integer"
                },
                "sort_order": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "tonnage": {
                    "description": "吨位",
                    "type": "number"
                },
                "width": {
                    "description": "船体宽度（米）",
                    "type": "number"
                }
            }
        },
...
    }
}